Add DNS CNAME record for studio auth
What we're doing here is claiming a lock for a set of previously
registered OAuth apps on GitHub. The difficulty is that the app callback
URI cannot be set programmtically (nor can a new app be created). The
idea here is that we'll CNAME an app's callback URI e.g.
auth.ci-1.storyscript-ci.com to the new env's auth server.