Use github oauth apps as locks
There was a problem with the previous solution of trying to be clever
with DNS records that SameSite cookies would not be usable on
studio.<env>.storyscript-ci.com when the github redirect URI was
studio.<app>.storyscript-ci.com. As such, the pool has been changed to
be a 1:1 mapping of GitHub Oauth Apps to Environments.

The pool will now claim a free GitHub app and create an environment with
that name, so if a new environment is desired, a new app should be
created and a lockfile with its name, containing the client id and
secret should be placed in the pool.